Add favicons, apple touch, mstile, android, manifest, and browserconfig
Add all images in the correct sizes and references in the manifest.json and
browserconfig.xml files. For the time being, the manifest.json is saved as
manifest.txt otherwise Sling tries to create nodes out of the file on build.
Run the following cURL command to fix it:

curl -u admin:admin \
http://localhost:8080/content/manifest.txt \
-F":operation=move" \
-F":dest=/content/manifest.json"

Update the readme to document the current workaround.